backend/pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "FastAPI backend for a modular RAG chatbot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.1",
    "pydantic-settings>=2.3.1",
    "pypdf>=4.2.0",
    "python-dotenv>=1.0.1",
    "python-docx>=1.1.0",
    "langchain>=0.2.6",
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-openai>=0.1.8",
    "langchain-chroma>=0.1.2",
    "langchain-text-splitters>=0.0.2",
    "langchain-cohere>=0.1.6",
    "chromadb>=0.5.3",
    "sse-starlette>=2.0.0",
    "orjson>=3.10.3",
    "aiofiles>=23.2.1",
    "python-multipart>=0.0.9",
    "tenacity>=8.2.3",
    "httpx>=0.27.0",
    "langsmith>=0.4.30"
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["db*", "logs*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "ruff>=0.4.7",
    "mypy>=1.8.0"
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"
